sql: Move the SCATTER tests into their own file
The scatter implementation was moved a while ago, so it makes sense to
me for the tests to move as well.

This is just code movement, it contains zero changes.

Release note: None
